+++ dash-vue2/src/App.vue
@@ -1,26 +1,7 @@
 <template>
-  <div id="app">
-    <div id="nav">
-      <router-link to="/">Home</router-link> |
-      <router-link to="/about">About</router-link>
-      
-    </div>
-    <router-view/>
-  </div>
+  <router-view/>
+  
 </template>
-<script>
-import {projectAuth} from '../src/firebase/config'
-import useLogin from '../src/firebase/useLogin'
-
-export default {
-  setup(){
-    const foo = useLogin
-    console.log(foo)
-    return foo;
-  },
-    
-}
-</script>
 
 <style>
 #app {
@@ -28,19 +9,8 @@ export default {
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
-  color: #2c3e50;
+  color: black;
+  
 }
 
-#nav {
-  padding: 30px;
-}
-
-#nav a {
-  font-weight: bold;
-  color: #2c3e50;
-}
-
-#nav a.router-link-exact-active {
-  color: #42b983;
-}
 </style>
