+++ dash-vue2/src/components/NavBar.vue
