+++ dash-vue2/src/components/ProjectDetail.vue
