+++ dash-vue2/src/views/CurriculumPage.vue
