+++ dash-vue2/src/views/Welcome.vue
