--- outreach-dash/.browserslistrc
@@ -1,3 +1,0 @@
-> 1%
-last 2 versions
-not dead
