--- outreach-dash/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "outreach-dash",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build"
-  },
-  "dependencies": {
-    "@firebase/app": "^0.6.13",
-    "@firebase/auth": "^0.16.1",
-    "@firebase/firestore": "^2.1.2",
-    "core-js": "^3.6.5",
-    "firebase": "^8.2.4",
-    "vue": "^3.0.5",
-    "vue-router": "^4.0.0-0"
-  },
-  "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-router": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/compiler-sfc": "^3.0.0"
-  }
-}
