--- outreach-dash/src/App.vue
@@ -1,16 +1,0 @@
-<template>
-  <router-view/>
-  
-</template>
-
-<style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: black;
-  
-}
-
-</style>
